requirements.txt
python-telegram-bot==12.5.1
requests==2.22.0
json5==0.8.5
bs4==0.0.1
urllib3==1.25.5